refactor: consolidate SuccessMessage to single typed implementation
Delete the legacy class-based `SuccessMessage.js` and update all
importers to use the typed functional component at
`components/messages/SuccessMessage.tsx`.

Note: pre-commit hook skipped due to pre-existing eslint errors in
FeatureImport.tsx (react-hooks/exhaustive-deps) unrelated to this change.
