+++ frontend/web/components/import-export/FeatureImport.tsx
@@ -23,7 +23,7 @@ import ErrorMessage from 'components/ErrorMessage'
 import InfoMessage from 'components/InfoMessage'
 import WarningMessage from 'components/WarningMessage'
 import FeatureListStore from 'common/stores/feature-list-store'
-import SuccessMessage from 'components/SuccessMessage'
+import SuccessMessage from 'components/messages/SuccessMessage'
 import TableSearchFilter from 'components/tables/TableSearchFilter'
 import Utils from 'common/utils/utils'
 import TableTagFilter from 'components/tables/TableTagFilter'
